feat: Add sample code for Station04
This commit adds sample code for Station04. The code includes a foreach loop that iterates over an array and breaks the loop when the value is equal to 5. It also includes a foreach loop in the Question class that calculates the sum of numbers in an array, skipping numbers that are multiples of 4 except for 4 itself.

New test output file "station04.xml" is also added.